requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
python-dotenv